refactor trace collector budget
